game_control/controllers/auto_battle_strategy.py
import json
import time
import random
import asyncio

class AutoBattleStrategy:
    def __init__(self, strategy_file):
        with open(strategy_file, 'r') as f:
            self.strategies = json.load(f)

    def get_strategy(self, battle_type):
        return self.strategies.get(battle_type, self.strategies['default'])

    async def execute_strategy(self, battle_type, game_engine):
        strategy = self.get_strategy(battle_type)
        for phase in strategy:
            if phase['type'] == 'action_sequence':
                await game_engine.perform_action_sequence(phase['actions'])
            elif phase['type'] == 'conditional':
                if await self.evaluate_condition(phase['condition'], game_engine):
                    await game_engine.perform_action_sequence(phase['actions'])
            elif phase['type'] == 'loop':
                start_time = time.time()
                while time.time() - start_time < phase['duration']:
                    await game_engine.perform_action_sequence(phase['actions'])
                    if 'break_condition' in phase and await self.evaluate_condition(phase['break_condition'], game_engine):
                        break
            elif phase['type'] == 'wait_for_object':
                await game_engine.wait_for_game_object(phase['template'], phase.get('timeout', 30))
            elif phase['type'] == 'ocr_action':
                text = await game_engine.ocr_utils.recognize_text(game_engine.capture_screen())
                if phase['text'] in text:
                    await game_engine.perform_action_sequence(phase['actions'])
            elif phase['type'] == 'random_action':
                action = random.choice(phase['actions'])
                await game_engine.perform_action_sequence([action])

    async def evaluate_condition(self, condition, game_engine):
        if condition['type'] == 'object_present':
            return bool(await game_engine.find_game_object(condition['template']))
        elif condition['type'] == 'game_state':
            return await game_engine.detect_game_state() == condition['state']
        elif condition['type'] == 'text_present':
            text = await game_engine.ocr_utils.recognize_text(game_engine.capture_screen())
            return condition['text'] in text
        elif condition['type'] == 'resource_check':
            resource_value = await game_engine.get_resource_value(condition['resource'])
            return resource_value >= condition['threshold']
        return False

if __name__ == "__main__":
    # 测试 AutoBattleStrategy 类
    strategy = AutoBattleStrategy('config/strategies.json')
    print(strategy.get_strategy('default'))
    # 注意：这里需要一个模拟的 game_engine 对象来完全测试 execute_strategy 方法